requirements.txt
Flask
Flask-Script
Flask-SQLAlchemy
Flask-Restless
